Additional experiments for i686 architecture

+++ properties_vs_compilers/ct/analyze.sh
@@ -2,8 +2,13 @@
 #
 # Run binsec on a set of binary files and check whether they enforce
 # secret-erasure or if they fail to scrub secret-data from memory.
-
-BINSEC_PATH=binsec_rel
+#
+# All results can be found in the `log` subdirectory. By default, if results are
+# found in `log`, they are not recomputed. Delete files in `log` to rerun binsec
+# and recompute the results.
+##
+LOGDIR="./log"
+BINSEC_PATH=binsec-rel
 
 # -------- Configure which binary files to analyze
 #
@@ -21,17 +26,18 @@ hacl_utility_all=$(echo "$hacl_utility_all" | sed 's/[^ ]* */hacl_utility\/&/g')
 ALL_PROGRAMS="${ct_select_all} ${ct_sort_all} ${openssl_utility_all} ${hacl_utility_all} ${tea_all}"
 
 # Add you new compiler version here
-CLANG_VERSIONS="clang-3.0_3.0 clang-3.9_3.9 clang_7.1.0 clang_9.0.1 clang_11.0.1"
-GCC_VERSIONS="gcc_5.4.0 gcc_6.2.0 gcc_7.2.0 gcc_8.3.0 gcc_10.2.0 arm-linux-gnueabi-gcc_10.3"
-ALL_COMPILERS="${CLANG_VERSIONS} ${GCC_VERSIONS}"
+CLANG_VERSIONS_x86="clang-3.0_3.0 clang-3.9_3.9 clang_7.1.0 clang_9.0.1 clang_11.0.1"
+GCC_VERSIONS_x86="gcc_5.4.0 gcc_6.2.0 gcc_7.2.0 gcc_8.3.0 gcc_10.2.0"
+ALL_COMPILERS_x86="${CLANG_VERSIONS_x86} ${GCC_VERSIONS_x86}"
+GCC_VERSIONS_ARM="arm-linux-gnueabi-gcc_10.3"
 
 # Add you new compiler options here
 OPTIMIZATIONS="O0 O1 O2 O3"
 OPTIMIZATIONS_CLANG="${OPTIMIZATIONS} O3-cmov"
 OPTIMIZATIONS_GCC="O0 O0-cmov O1 O2 O3 O3-cmov"
+ARCH="i386 i686"
 
 # -------- Configure debug options
-LOGDIR="./log"
 TRACEDIR="/tmp/SMTDIR/trace/"
 
 # Enables debug with 'y' and disable with 'n'
@@ -51,6 +57,10 @@ else
     ADD_DEBUG=" -relse-debug-level 0"
 fi
 
+for dir in ct-select ct-sort hacl_utility openssl_utility tea; do
+    mkdir -p ${PWD}/log/${dir}
+done
+
 # [run binary_file] Run binsec on the specified [binary_file]
 run() {
     NAME="${1}"
@@ -65,14 +75,21 @@ run() {
     -relse-memory-type row-map -relse-property ct -relse-dedup 1 \
     -relse-fp instr -relse-leak-info instr bin/${NAME} ${ADD_DEBUG}"
 
-    echo "${CMD}" > "${LOGFILE}"
-    eval "${CMD}" >> "${LOGFILE}" 2>&1
-    RES=$?
-    if [ ${RES} -eq 0 ]; then
+    if ! [ -f "${LOGFILE}" ]; then
+        echo "${CMD}" > "${LOGFILE}"
+        eval "${CMD}" >> "${LOGFILE}" 2>&1
+    fi
+
+    if grep "Result:	Secure" "${LOGFILE}" >> /dev/null; then
         RESULT="SECURE"
-    elif [ ${RES} -eq 7 ]; then
-        RESULT="INSECURE"
-    elif [ ${RES} -eq 8 ]; then
+    elif grep "Result:	Insecure" "${LOGFILE}"  >> /dev/null; then
+        RESULT=""
+        if grep "\[relse:result\] \[Insecurity\]\[Violation\] Insecure memory access" "${LOGFILE}" >> /dev/null; then
+            RESULT="MEM"
+        elif grep "\[relse:result\] \[Insecurity\]\[Violation\] Insecure jump" "${LOGFILE}" >> /dev/null; then
+             RESULT="${RESULT}CF"
+        fi
+    elif grep "Result:  Unknown" "${LOGFILE}" >> /dev/null; then
         RESULT="UNKNOWN"
     else
         RESULT="ERROR"
@@ -108,37 +125,26 @@ unit_tests() {
     check "./ct-select/ct_select_v1_O3_gcc_8.3.0" "SECURE"
     check "./ct-select/ct_select_v1_O0-cmov_gcc_8.3.0" "SECURE"
     check "./ct-select/ct_select_v1_O0_clang_7.1.0" "SECURE"
-    check "./ct-select/ct_select_v1_O3_clang_7.1.0" "INSECURE"
-    check "./ct-select/ct_select_v1_O3-cmov_clang_7.1.0" "INSECURE"
+    check "./ct-select/ct_select_v1_O3_clang_7.1.0" "CF"
+    check "./ct-select/ct_select_v1_O3-cmov_clang_7.1.0" "CF"
 
-    check "./ct-sort/sort_O0_gcc_8.3.0" "INSECURE"
+    check "./ct-sort/sort_O0_gcc_8.3.0" "CF"
     check "./ct-sort/sort_O3_gcc_8.3.0" "SECURE"
-    check "./ct-sort/sort_O0-cmov_gcc_8.3.0" "INSECURE"
+    check "./ct-sort/sort_O0-cmov_gcc_8.3.0" "CF"
     check "./ct-sort/sort_O3-cmov_gcc_8.3.0" "SECURE"
     check "./ct-sort/sort_O0_clang_7.1.0" "SECURE"
     check "./ct-sort/sort_O3_clang_7.1.0" "SECURE"
     check "./ct-sort/sort_O3-cmov_clang_7.1.0" "SECURE"
 
-    check "./ct-select/naive_select_O3-cmov_clang_7.1.0" "INSECURE"
-    check "./ct-select/naive_select_O3-cmov_gcc_8.3.0" "INSECURE"
-    check "./ct-select/naive_select_O0_arm-linux-gnueabi-gcc_10.3" "INSECURE"
-    check "./ct-select/naive_select_O0-cmov_arm-linux-gnueabi-gcc_10.3" "INSECURE"
+    check "./ct-select/naive_select_O3-cmov_clang_7.1.0" "CF"
+    check "./ct-select/naive_select_O3-cmov_gcc_8.3.0" "CF"
+    check "./ct-select/naive_select_O0_arm-linux-gnueabi-gcc_10.3" "CF"
+    check "./ct-select/naive_select_O0-cmov_arm-linux-gnueabi-gcc_10.3" "CF"
     check "./ct-select/naive_select_O3_arm-linux-gnueabi-gcc_10.3" "SECURE"
-    check "./ct-select/naive_select_O3-cmov_arm-linux-gnueabi-gcc_10.3" "INSECURE"
+    check "./ct-select/naive_select_O3-cmov_arm-linux-gnueabi-gcc_10.3" "CF"
 
     printf "\n Unit tests passed !\n\n"
 }
-
-
-# Generate latex tables
-# read -ar clang_array <<< "${CLANG_VERSIONS}"
-# (read -a ${CLANG_VERSIONS})
-clang_array=(${CLANG_VERSIONS})
-last_clang_compiler="${clang_array[-1]}"
-gcc_array=(${GCC_VERSIONS})
-last_gcc_compiler="${gcc_array[-1]}"
-optimizations_array=(${OPTIMIZATIONS})
-last_optimization="${optimizations_array[-1]}"
 
 print_results_latex () {
     programs="$1"
@@ -148,12 +154,16 @@ print_results_latex () {
     run "${NAME}"
     if [ ${RESULT} = "SECURE" ]; then
         echo -n "\ccmark{}"
-    elif [ ${RESULT} = "INSECURE" ]; then
-        echo -n "\cxmark{}"
+    elif [ ${RESULT} = "CF" ]; then
+        echo -n "\textsc{\textcolor{red}{c}}"
+    elif [ ${RESULT} = "MEM" ]; then
+        echo -n "\textsc{\textcolor{red}{m}}"
+    elif [ ${RESULT} = "MEMCF" ]; then
+        echo -n "\textsc{\textcolor{red}{b}}"
     elif [ ${RESULT} = "UNKNOWN" ]; then
         echo -n "\csmark{}"
     else
-        echo -n "\textsc{na}"
+        echo -n "-"
     fi
 }
 
@@ -178,6 +188,144 @@ generate_latex_table () {
                 else
                     echo -n " & "
                 fi
+                if [ "${4}" != "" ]; then
+                    option="${option}-${4}"
+                fi
+                print_results_latex "${program}" "${compiler}" "${option}"
+            done
+        done
+        echo "\\\\"
+    done
+}
+
+# Generate final clang table for the paper
+generate_latex_table_clang () {
+    for program in ${1}
+    do
+        first="t"
+        pp_program_name_to_latex "${program}"
+        for compiler in clang-3.0_3.0 clang-3.9_3.9
+        do
+            for option in ${OPTIMIZATIONS_CLANG}
+            do
+                if [ "${first}" = "t" ]; then
+                    first="f"
+                else
+                    echo -n " & "
+                fi
+                print_results_latex "${program}" "${compiler}" "${option}-i686"
+            done
+        done
+        for compiler in clang_7.1.0 clang_9.0.1
+        do
+            for arch in i386 i686; do
+                for option in ${OPTIMIZATIONS_CLANG}
+                do
+                    if [ "${first}" = "t" ]; then
+                        first="f"
+                    else
+                        echo -n " & "
+                    fi
+                    print_results_latex "${program}" "${compiler}" "${option}-${arch}"
+                done
+            done
+        done
+        echo "\\\\"
+    done
+}
+
+# Final gcc table for the paper
+generate_latex_table_gcc () {
+    for program in ${1}
+    do
+        first="t"
+        pp_program_name_to_latex "${program}"
+        for compiler in "gcc_10.2.0"
+        do
+            for arch in i386 i686; do
+                for option in ${OPTIMIZATIONS_GCC}
+                do
+                    if [ "${first}" = "t" ]; then
+                        first="f"
+                    else
+                        echo -n " & "
+                    fi
+                    print_results_latex "${program}" "${compiler}" "${option}-${arch}"
+                done
+            done
+        done
+        for compiler in "arm-linux-gnueabi-gcc_10.3"
+        do
+            for option in ${OPTIMIZATIONS_GCC}
+            do
+                if [ "${first}" = "t" ]; then
+                    first="f"
+                else
+                    echo -n " & "
+                fi
+                print_results_latex "${program}" "${compiler}" "${option}"
+            done
+        done
+        echo "\\\\"
+    done
+}
+
+# Ugly ad-hoc function to generate final table for the paper
+generate_latex_table_clang_gcc () {
+    programs="${ct_select_all} ${ct_sort_all} ${openssl_utility_all%% *} ${hacl_utility_all%% *} ${tea_all%% *}"
+    for program in  ${programs}
+    do
+        first="t"
+        pp_program_name_to_latex "${program}"
+        for compiler in clang-3.0_3.0 clang-3.9_3.9
+        do
+            for option in ${OPTIMIZATIONS_CLANG}
+            do
+                if [ "${first}" = "t" ]; then
+                    first="f"
+                else
+                    echo -n " & "
+                fi
+                print_results_latex "${program}" "${compiler}" "${option}-i686"
+            done
+        done
+        for compiler in clang_7.1.0 clang_9.0.1
+        do
+            for arch in i386 i686; do
+                for option in ${OPTIMIZATIONS_CLANG}
+                do
+                    if [ "${first}" = "t" ]; then
+                        first="f"
+                    else
+                        echo -n " & "
+                    fi
+                    print_results_latex "${program}" "${compiler}" "${option}-${arch}"
+                done
+            done
+        done
+        for compiler in "gcc_10.2.0"
+        do
+            for arch in i386 i686; do
+                for option in ${OPTIMIZATIONS_GCC}
+                do
+                    if [ "${first}" = "t" ]; then
+                        first="f"
+                    else
+                        echo -n " & "
+                    fi
+                    print_results_latex "${program}" "${compiler}" "${option}-${arch}"
+                done
+            done
+        done
+        for compiler in "arm-linux-gnueabi-gcc_10.3"
+        do
+            for option in ${OPTIMIZATIONS_GCC}
+            do
+                if [ "${first}" = "t" ]; then
+                    first="f"
+                else
+                    echo -n " & "
+                fi
                 print_results_latex "${program}" "${compiler}" "${option}"
             done
         done
@@ -191,9 +339,12 @@ pp_results () {
     do
         for compiler in ${2}
         do
-            for optimization in ${3}
+            for option in ${3}
             do
-                NAME="${program}_${optimization}_${compiler}"
+                if [ "${4}" != "" ]; then
+                    option="${option}-${4}"
+                fi
+                NAME="${program}_${option}_${compiler}"
                 run "${NAME}"
                 echo "$NAME: ${RESULT}"
             done
@@ -202,20 +353,39 @@ pp_results () {
 }
 
 wrong_usage() {
-    echo "Usage: analyze.sh [ test | latex | all | compiler <compiler_name> | program <program_path> ]"
+    echo "Usage: analyze.sh [ test | latex <output_dir> | all | compiler <compiler_name> | program <program_path> ]"
     exit 1
 }
 
 if [ "${1}" = "test" ]; then
     unit_tests
 elif [ "${1}" = "latex" ]; then
-    echo "CLANG"
-    generate_latex_table "${ALL_PROGRAMS}" "${CLANG_VERSIONS}" "${OPTIMIZATIONS_CLANG}"
-    echo "GCC"
-    generate_latex_table "${ALL_PROGRAMS}" "${GCC_VERSIONS}" "${OPTIMIZATIONS_GCC}"
+    if [ "${2}" = "" ]; then
+        wrong_usage;
+    else
+        # echo "% -- CLANG-i386" > ${2}/clang-i386.tex
+        # generate_latex_table "${ALL_PROGRAMS}" "${CLANG_VERSIONS_x86}" "${OPTIMIZATIONS_CLANG}" "i386"  >> ${2}/clang-i386.tex
+        # echo "% -- GCC-i386" > ${2}/gcc-i386.tex
+        # generate_latex_table "${ALL_PROGRAMS}" "${GCC_VERSIONS_x86}" "${OPTIMIZATIONS_GCC}" "i386" >> ${2}/gcc-i386.tex
+        # echo "% -- CLANG-i686" > ${2}/clang-i686.tex
+        # generate_latex_table "${ALL_PROGRAMS}" "${CLANG_VERSIONS_x86}" "${OPTIMIZATIONS_CLANG}" "i686" >> ${2}/clang-i686.tex
+        # echo "% -- GCC-i686" > ${2}/gcc-i686.tex
+        # generate_latex_table "${ALL_PROGRAMS}" "${GCC_VERSIONS_x86}" "${OPTIMIZATIONS_GCC}" "i686" >> ${2}/gcc-i686.tex
+        # echo "% -- GCC-arm" > ${2}/gcc-arm.tex
+        # generate_latex_table "${ALL_PROGRAMS}" "${GCC_VERSIONS_ARM}" "${OPTIMIZATIONS_GCC}" "" >> ${2}/gcc-arm.tex
+        # echo "% -- CLANG" > ${2}/clang.tex
+        # generate_latex_table_clang "${ALL_PROGRAMS}" >> ${2}/clang.tex
+        # echo "% -- GCC" > ${2}/gcc.tex
+        # generate_latex_table_gcc "${ALL_PROGRAMS}" >> ${2}/gcc.tex
+        echo "% -- CLANG-GCC" > ${2}/clang-gcc.tex
+        generate_latex_table_clang_gcc >> ${2}/clang-gcc.tex
+    fi
 elif [ "${1}" = "all" ]; then
-    pp_results "${ALL_PROGRAMS}" "${CLANG_VERSIONS}" "${OPTIMIZATIONS_CLANG}"
-    pp_results "${ALL_PROGRAMS}" "${GCC_VERSIONS}" "${OPTIMIZATIONS_GCC}"
+    pp_results "${ALL_PROGRAMS}" "${CLANG_VERSIONS_x86}" "${OPTIMIZATIONS_CLANG}" "i386"
+    pp_results "${ALL_PROGRAMS}" "${GCC_VERSIONS_x86}" "${OPTIMIZATIONS_GCC}" "i386"
+    pp_results "${ALL_PROGRAMS}" "${CLANG_VERSIONS_x86}" "${OPTIMIZATIONS_CLANG}" "i686"
+    pp_results "${ALL_PROGRAMS}" "${GCC_VERSIONS_x86}" "${OPTIMIZATIONS_GCC}" "i686"
+    pp_results "${ALL_PROGRAMS}" "${GCC_VERSIONS_ARM}" "${OPTIMIZATIONS_GCC}" ""
 elif [ "${1}" = "compiler" ]; then
     if [ "${2}" = "" ]; then
         wrong_usage;
